#434 Move duplicated onError(...) method to utility class
