+++ rule/asynchttpclient/src/main/java/io/opentracing/contrib/specialagent/rule/asynchttpclient/TracingAsyncHandler.java
@@ -15,10 +15,9 @@
 
 package io.opentracing.contrib.specialagent.rule.asynchttpclient;
 
+import io.opentracing.contrib.specialagent.SpanUtil;
 import java.net.InetSocketAddress;
-import java.util.HashMap;
 import java.util.List;
-import java.util.Map;
 
 import javax.net.ssl.SSLSession;
 
@@ -73,7 +72,7 @@ public class TracingAsyncHandler implements AsyncHandler<Object> {
       handler.onThrowable(t);
     }
     finally {
-      onError(t, span);
+      SpanUtil.onError(t, span);
     }
   }
 
@@ -155,16 +154,4 @@ public class TracingAsyncHandler implements AsyncHandler<Object> {
     handler.onRetry();
   }
 
-  private static void onError(final Throwable throwable, final Span span) {
-    Tags.ERROR.set(span, Boolean.TRUE);
-    if (throwable != null)
-      span.log(errorLogs(throwable));
-  }
-
-  private static Map<String,Object> errorLogs(final Throwable throwable) {
-    final Map<String,Object> errorLogs = new HashMap<>(2);
-    errorLogs.put("event", Tags.ERROR.getKey());
-    errorLogs.put("error.object", throwable);
-    return errorLogs;
-  }
 }
