+++ rule/aws-sdk-2/src/main/java/io/opentracing/contrib/specialagent/rule/aws/sdk2/TracingExecutionInterceptor.java
@@ -15,8 +15,7 @@
 
 package io.opentracing.contrib.specialagent.rule.aws.sdk2;
 
-import java.util.HashMap;
-import java.util.Map;
+import io.opentracing.contrib.specialagent.SpanUtil;
 
 import io.opentracing.Span;
 import io.opentracing.tag.Tags;
@@ -75,15 +74,8 @@ public class TracingExecutionInterceptor implements ExecutionInterceptor {
       return;
 
     executionAttributes.putAttribute(SPAN_ATTRIBUTE, null);
-    Tags.ERROR.set(span, Boolean.TRUE);
-    span.log(errorLogs(context.exception()));
+    SpanUtil.onError(context.exception(), span);
     span.finish();
   }
 
-  private static Map<String,Object> errorLogs(final Throwable ex) {
-    Map<String,Object> errorLogs = new HashMap<>(2);
-    errorLogs.put("event", Tags.ERROR.getKey());
-    errorLogs.put("error.object", ex);
-    return errorLogs;
-  }
 }
