+++ rule/netty/src/main/java/io/opentracing/contrib/specialagent/rule/netty/TracingServerChannelInboundHandlerAdapter.java
@@ -15,8 +15,7 @@
 
 package io.opentracing.contrib.specialagent.rule.netty;
 
-import java.util.HashMap;
-import java.util.Map;
+import io.opentracing.contrib.specialagent.SpanUtil;
 
 import io.netty.channel.ChannelHandlerContext;
 import io.netty.channel.ChannelInboundHandlerAdapter;
@@ -62,23 +61,11 @@ public class TracingServerChannelInboundHandlerAdapter extends ChannelInboundHan
         handlerContext.fireChannelRead(message);
       }
       catch (final Throwable throwable) {
-        onError(throwable, span);
+        SpanUtil.onError(throwable, span);
         span.finish();
         throw throwable;
       }
     }
   }
 
-  public static void onError(final Throwable t, final Span span) {
-    Tags.ERROR.set(span, Boolean.TRUE);
-    if (t != null)
-      span.log(errorLogs(t));
-  }
-
-  private static Map<String,Object> errorLogs(final Throwable t) {
-    final Map<String,Object> errorLogs = new HashMap<>(2);
-    errorLogs.put("event", Tags.ERROR.getKey());
-    errorLogs.put("error.object", t);
-    return errorLogs;
-  }
 }
