+++ rule/pulsar-client/src/main/java/io/opentracing/contrib/specialagent/rule/pulsar/client/PulsarClientAgentIntercept.java
@@ -16,8 +16,7 @@
 package io.opentracing.contrib.specialagent.rule.pulsar.client;
 
 import io.opentracing.contrib.specialagent.LocalSpanContext;
-import java.util.HashMap;
-import java.util.Map;
+import io.opentracing.contrib.specialagent.SpanUtil;
 import java.util.concurrent.CompletableFuture;
 
 import org.apache.pulsar.client.api.Consumer;
@@ -115,7 +114,7 @@ public class PulsarClientAgentIntercept {
     contextHolder.remove();
 
     if (thrown != null) {
-      onError(thrown, span);
+      SpanUtil.onError(thrown, span);
       span.finish();
       return returned;
     }
@@ -124,22 +123,10 @@ public class PulsarClientAgentIntercept {
       span.finish();
       return messageId;
     }).exceptionally(throwable -> {
-      onError(throwable, span);
+      SpanUtil.onError(throwable, span);
       span.finish();
       return null;
     });
   }
 
-  private static void onError(final Throwable t, final Span span) {
-    Tags.ERROR.set(span, Boolean.TRUE);
-    if (t != null)
-      span.log(errorLogs(t));
-  }
-
-  private static Map<String,Object> errorLogs(final Throwable t) {
-    final Map<String,Object> errorLogs = new HashMap<>(2);
-    errorLogs.put("event", Tags.ERROR.getKey());
-    errorLogs.put("error.object", t);
-    return errorLogs;
-  }
 }
