+++ rule/rabbitmq-client/src/main/java/io/opentracing/contrib/specialagent/rule/rabbitmq/client/RabbitMQAgentIntercept.java
@@ -18,8 +18,7 @@ package io.opentracing.contrib.specialagent.rule.rabbitmq.client;
 import static io.opentracing.contrib.rabbitmq.TracingUtils.*;
 
 import io.opentracing.contrib.specialagent.LocalSpanContext;
-import java.util.HashMap;
-import java.util.Map;
+import io.opentracing.contrib.specialagent.SpanUtil;
 
 import com.rabbitmq.client.AMQP;
 import com.rabbitmq.client.Consumer;
@@ -31,7 +30,6 @@ import io.opentracing.Tracer;
 import io.opentracing.contrib.common.WrapperProxy;
 import io.opentracing.contrib.rabbitmq.TracingConsumer;
 import io.opentracing.contrib.rabbitmq.TracingUtils;
-import io.opentracing.tag.Tags;
 import io.opentracing.util.GlobalTracer;
 
 public class RabbitMQAgentIntercept {
@@ -40,7 +38,7 @@ public class RabbitMQAgentIntercept {
   public static void exitGet(final Object response, final Object queue, final Throwable thrown) {
     final Span span = TracingUtils.buildChildSpan(((GetResponse)response).getProps(), (String)queue, GlobalTracer.get());
     if (thrown != null)
-      captureException(span, thrown);
+      SpanUtil.onError(thrown, span);
 
     span.finish();
   }
@@ -70,17 +68,10 @@ public class RabbitMQAgentIntercept {
       return;
 
     if (thrown != null)
-      captureException(context.getSpan(), thrown);
+      SpanUtil.onError(thrown, context.getSpan());
 
     context.closeAndFinish();
     contextHolder.remove();
   }
 
-  private static void captureException(final Span span, final Throwable thrown) {
-    final Map<String,Object> exceptionLogs = new HashMap<>();
-    exceptionLogs.put("event", Tags.ERROR.getKey());
-    exceptionLogs.put("error.object", thrown);
-    span.log(exceptionLogs);
-    Tags.ERROR.set(span, true);
-  }
 }
