+++ rule/spring-kafka/src/main/java/io/opentracing/contrib/specialagent/rule/spring/kafka/SpringKafkaAgentIntercept.java
@@ -16,8 +16,7 @@
 package io.opentracing.contrib.specialagent.rule.spring.kafka;
 
 import io.opentracing.contrib.specialagent.LocalSpanContext;
-import java.util.HashMap;
-import java.util.Map;
+import io.opentracing.contrib.specialagent.SpanUtil;
 
 import org.apache.kafka.clients.consumer.ConsumerRecord;
 
@@ -62,7 +61,7 @@ public class SpringKafkaAgentIntercept {
     if (context != null) {
       if (context.decrementAndGet() == 0) {
         if (thrown != null) {
-          captureException(context.getSpan(), thrown);
+          SpanUtil.onError(thrown, context.getSpan());
         }
         context.closeAndFinish();
         contextHolder.remove();
@@ -70,11 +69,4 @@ public class SpringKafkaAgentIntercept {
     }
   }
 
-  private static void captureException(final Span span, final Throwable t) {
-    final Map<String,Object> exceptionLogs = new HashMap<>();
-    exceptionLogs.put("event", Tags.ERROR.getKey());
-    exceptionLogs.put("error.object", t);
-    span.log(exceptionLogs);
-    Tags.ERROR.set(span, true);
-  }
 }
