+++ rule/spring-scheduling/src/main/java/io/opentracing/contrib/specialagent/rule/spring/scheduling/SpringSchedulingAgentIntercept.java
@@ -16,8 +16,7 @@
 package io.opentracing.contrib.specialagent.rule.spring.scheduling;
 
 import io.opentracing.contrib.specialagent.LocalSpanContext;
-import java.util.HashMap;
-import java.util.Map;
+import io.opentracing.contrib.specialagent.SpanUtil;
 
 import org.aopalliance.intercept.MethodInvocation;
 import org.springframework.scheduling.support.ScheduledMethodRunnable;
@@ -53,18 +52,10 @@ public class SpringSchedulingAgentIntercept {
       return;
 
     if (thrown != null)
-      captureException(context.getSpan(), thrown);
+      SpanUtil.onError(thrown, context.getSpan());
 
     context.closeAndFinish();
     contextHolder.remove();
-  }
-
-  static void captureException(final Span span, final Throwable t) {
-    final Map<String,Object> exceptionLogs = new HashMap<>();
-    exceptionLogs.put("event", Tags.ERROR.getKey());
-    exceptionLogs.put("error.object", t);
-    span.log(exceptionLogs);
-    Tags.ERROR.set(span, true);
   }
 
   public static Object invoke(final Object arg) {
