+++ rule/spring-web-4.0/src/main/java/io/opentracing/contrib/specialagent/rule/spring/web40/copied/RestTemplateSpanDecorator.java
@@ -14,10 +14,9 @@
 package io.opentracing.contrib.specialagent.rule.spring.web40.copied;
 
 import io.opentracing.Span;
+import io.opentracing.contrib.specialagent.SpanUtil;
 import io.opentracing.tag.Tags;
 import java.io.IOException;
-import java.util.HashMap;
-import java.util.Map;
 import org.apache.commons.logging.Log;
 import org.apache.commons.logging.LogFactory;
 import org.springframework.http.HttpRequest;
@@ -96,15 +95,8 @@ public interface RestTemplateSpanDecorator {
 
     @Override
     public void onError(HttpRequest httpRequest, Throwable ex, Span span) {
-      Tags.ERROR.set(span, Boolean.TRUE);
-      span.log(errorLogs(ex));
+      SpanUtil.onError(ex, span);
     }
 
-    public static Map<String, Object> errorLogs(Throwable ex) {
-      Map<String, Object> errorLogs = new HashMap<>(2);
-      errorLogs.put("event", Tags.ERROR.getKey());
-      errorLogs.put("error.object", ex);
-      return errorLogs;
-    }
   }
 }
