+++ rule/spring-web-4/src/main/java/io/opentracing/contrib/specialagent/rule/spring/web4/copied/TracingListenableFutureCallback.java
@@ -16,10 +16,9 @@ package io.opentracing.contrib.specialagent.rule.spring.web4.copied;
 
 import io.opentracing.Scope;
 import io.opentracing.Span;
+import io.opentracing.contrib.specialagent.SpanUtil;
 import io.opentracing.tag.Tags;
 import io.opentracing.util.GlobalTracer;
-import java.util.HashMap;
-import java.util.Map;
 import org.springframework.http.ResponseEntity;
 import org.springframework.util.concurrent.ListenableFutureCallback;
 
@@ -38,7 +37,7 @@ public class TracingListenableFutureCallback implements ListenableFutureCallback
   @Override
   public void onFailure(Throwable ex) {
     if (finishSpan) {
-      captureException(span, ex);
+      SpanUtil.onError(ex, span);
       span.finish();
     }
     if (callback != null) {
@@ -66,11 +65,4 @@ public class TracingListenableFutureCallback implements ListenableFutureCallback
 
   }
 
-  public static void captureException(final Span span, final Throwable t) {
-    final Map<String, Object> exceptionLogs = new HashMap<>();
-    exceptionLogs.put("event", Tags.ERROR.getKey());
-    exceptionLogs.put("error.object", t);
-    span.log(exceptionLogs);
-    Tags.ERROR.set(span, true);
-  }
 }
