+++ rule/spring-web-5/src/main/java/io/opentracing/contrib/specialagent/rule/spring/web5/copied/RestTemplateSpanDecorator.java
@@ -13,9 +13,8 @@
  */
 package io.opentracing.contrib.specialagent.rule.spring.web5.copied;
 
+import io.opentracing.contrib.specialagent.SpanUtil;
 import java.io.IOException;
-import java.util.HashMap;
-import java.util.Map;
 
 import org.apache.commons.logging.Log;
 import org.apache.commons.logging.LogFactory;
@@ -98,15 +97,8 @@ public interface RestTemplateSpanDecorator {
 
     @Override
     public void onError(HttpRequest httpRequest, Throwable ex, Span span) {
-      Tags.ERROR.set(span, Boolean.TRUE);
-      span.log(errorLogs(ex));
+      SpanUtil.onError(ex, span);
     }
 
-    public static Map<String, Object> errorLogs(Throwable ex) {
-      Map<String, Object> errorLogs = new HashMap<>(2);
-      errorLogs.put("event", Tags.ERROR.getKey());
-      errorLogs.put("error.object", ex);
-      return errorLogs;
-    }
   }
 }
