+++ rule/spymemcached/src/main/java/io/opentracing/contrib/specialagent/rule/spymemcached/TracingOperationCallback.java
@@ -15,11 +15,11 @@
 
 package io.opentracing.contrib.specialagent.rule.spymemcached;
 
+import io.opentracing.contrib.specialagent.SpanUtil;
 import java.util.HashMap;
 import java.util.Map;
 
 import io.opentracing.Span;
-import io.opentracing.tag.Tags;
 import net.spy.memcached.ops.OperationCallback;
 import net.spy.memcached.ops.OperationStatus;
 
@@ -51,11 +51,7 @@ public class TracingOperationCallback implements OperationCallback {
   }
 
   void onError(final Throwable thrown) {
-    Tags.ERROR.set(span, Boolean.TRUE);
-    final HashMap<String,Object> errorLogs = new HashMap<>(2);
-    errorLogs.put("event", Tags.ERROR.getKey());
-    errorLogs.put("error.object", thrown);
-    span.log(errorLogs);
+    SpanUtil.onError(thrown, span);
     span.finish();
   }
 }
